// packages/framework/src/InputTouch.type.ts
export type InputTouch = {
  CLIENT_X: number;
  CLIENT_Y: number;
  DOWN_INITIAL_CLIENT_X: number;
  DOWN_INITIAL_CLIENT_Y: number;
  STRETCH_X: number;
  STRETCH_Y: number;
  IN_BOUNDS: number;
  PRESSURE: number;
  RELATIVE_X: number;
  RELATIVE_Y: number;
  STRETCH_VECTOR_X: number;
  STRETCH_VECTOR_Y: number;
  VECTOR_X: number;
  VECTOR_Y: number;
};